+++ app/templates/config.html
@@ -10,7 +10,7 @@
                 <h4 class="mb-0"><i class="bi bi-gear me-2"></i>Terminal Connect Configuration</h4>
             </div>
             <div class="card-body">
-                <form method="POST" action="{{ url_for('main.config') }}">
+                <form method="POST" action="{{ url_for('config.config') }}">
                     <div class="mb-4">
                         <label for="environment" class="form-label">Environment</label>
                         <select class="form-select" id="environment" name="environment" required>
