+++ app/templates/linked_refund.html
@@ -4,54 +4,32 @@
 
 {% block content %}
 <div class="row justify-content-center">
-    <div class="col-md-6">
+    <div class="col-md-8">
         <div class="card">
             <div class="card-header">
-                <h4 class="mb-0"><i class="bi bi-link me-2"></i>Process Linked Refund</h4>
+                <h4 class="mb-0"><i class="bi bi-arrow-return-left me-2"></i>Linked Refund</h4>
             </div>
             <div class="card-body">
-                <form method="POST" action="{{ url_for('main.linked_refund') }}">
+                <form method="POST" action="{{ url_for('refunds.linked_refund') }}">
                     <div class="mb-4">
                         <label for="amount" class="form-label">Amount</label>
                         <div class="input-group">
                             <span class="input-group-text"><i class="bi bi-currency-dollar"></i></span>
-                            <input type="number" 
-                                   class="form-control" 
-                                   id="amount" 
-                                   name="amount" 
-                                   step="0.01" 
-                                   min="0.01" 
-                                   pattern="^\d*\.?\d{0,2}$"
-                                   placeholder="0.00"
-                                   required>
+                            <input type="number" class="form-control" id="amount" name="amount" step="0.01" min="0.01" required>
                         </div>
                     </div>
                     <div class="mb-4">
                         <label for="merchant_reference" class="form-label">Merchant Reference</label>
                         <div class="input-group">
-                            <span class="input-group-text"><i class="bi bi-hash"></i></span>
-                            <input type="text" class="form-control" id="merchant_reference" name="merchant_reference" required>
-                        </div>
-                        <div class="form-text">
-                            <i class="bi bi-info-circle me-1"></i>
-                            A unique reference for this refund
+                            <span class="input-group-text"><i class="bi bi-tag"></i></span>
+                            <input type="text" class="form-control" id="merchant_reference" name="merchant_reference" value="{{ default_merchant_reference }}" required>
                         </div>
                     </div>
                     <div class="mb-4">
-                        <label for="parent_intent_id" class="form-label">Parent Intent ID</label>
+                        <label for="original_merchant_reference" class="form-label">Original Sale Reference</label>
                         <div class="input-group">
-                            <span class="input-group-text"><i class="bi bi-key"></i></span>
-                            <input type="text" 
-                                   class="form-control" 
-                                   id="parent_intent_id" 
-                                   name="parent_intent_id" 
-                                   pattern="^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
-                                   placeholder="xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx"
-                                   required>
-                        </div>
-                        <div class="form-text">
-                            <i class="bi bi-info-circle me-1"></i>
-                            The Intent ID of the original sale transaction (must be a valid UUID v4)
+                            <span class="input-group-text"><i class="bi bi-link"></i></span>
+                            <input type="text" class="form-control" id="original_merchant_reference" name="original_merchant_reference" required>
                         </div>
                     </div>
                     <div class="mb-4">
@@ -77,7 +55,7 @@
                     </div>
                     <div class="d-grid">
                         <button type="submit" class="btn btn-primary">
-                            <i class="bi bi-link me-2"></i>Process Linked Refund
+                            <i class="bi bi-arrow-return-left me-2"></i>Process Refund
                         </button>
                     </div>
                 </form>
