+++ app/templates/sale.html
@@ -4,42 +4,30 @@
 
 {% block content %}
 <div class="row justify-content-center">
-    <div class="col-md-6">
+    <div class="col-md-8">
         <div class="card">
             <div class="card-header">
-                <h4 class="mb-0"><i class="bi bi-cart me-2"></i>Process Sale</h4>
+                <h4 class="mb-0"><i class="bi bi-cart me-2"></i>Sale</h4>
             </div>
             <div class="card-body">
-                <form method="POST" action="{{ url_for('main.sale') }}">
+                <form method="POST" action="{{ url_for('sales.sale') }}">
                     <div class="mb-4">
                         <label for="amount" class="form-label">Amount</label>
                         <div class="input-group">
                             <span class="input-group-text"><i class="bi bi-currency-dollar"></i></span>
-                            <input type="number" 
-                                   class="form-control" 
-                                   id="amount" 
-                                   name="amount" 
-                                   step="0.01" 
-                                   min="0.01" 
-                                   pattern="^\d*\.?\d{0,2}$"
-                                   placeholder="0.00"
-                                   required>
+                            <input type="number" class="form-control" id="amount" name="amount" step="0.01" min="0.01" required>
                         </div>
                     </div>
                     <div class="mb-4">
                         <label for="merchant_reference" class="form-label">Merchant Reference</label>
                         <div class="input-group">
-                            <span class="input-group-text"><i class="bi bi-hash"></i></span>
-                            <input type="text" class="form-control" id="merchant_reference" name="merchant_reference" required>
-                        </div>
-                        <div class="form-text">
-                            <i class="bi bi-info-circle me-1"></i>
-                            A unique reference for this transaction
+                            <span class="input-group-text"><i class="bi bi-tag"></i></span>
+                            <input type="text" class="form-control" id="merchant_reference" name="merchant_reference" value="{{ default_merchant_reference }}" required>
                         </div>
                     </div>
                     <div class="d-grid">
                         <button type="submit" class="btn btn-primary">
-                            <i class="bi bi-cart-check me-2"></i>Process Sale
+                            <i class="bi bi-credit-card me-2"></i>Process Sale
                         </button>
                     </div>
                 </form>
